Add Respoke-SDK header to requests
This patch adds a "Respoke-SDK" header to all requests, to help
Respoke get more insight into which client libraries are being used
and what environments they are being run in. The header contains
The version of the respoke-php client library, the os and version,
and the version of php being used.

Related to MER-4340.